% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{blockr.stats-package}
\alias{blockr.stats}
\alias{blockr.stats-package}
\title{blockr.stats: Statistical Blocks for blockr}
\description{
Extends 'blockr.core' with no-code statistical blocks: the applied-stats-basics spine (describe, compare, associate, regress, nonparametric) plus an advanced tier (survival). Blocks fit models and tests and emit tidy data frames consumed by generic renderers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/blockr-org/blockr.stats}
}

}
\author{
\strong{Maintainer}: Christoph Sax \email{christoph@cynkra.com} (\href{https://orcid.org/0000-0002-7192-7044}{ORCID})

}
\keyword{internal}
