% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{blockr.stats-package}
\alias{blockr.stats}
\alias{blockr.stats-package}
\title{blockr.stats: Statistical Test Blocks for 'blockr'}
\description{
Extends 'blockr.core' with an interactive block for statistical hypothesis tests. Provides a single adaptive block supporting normality tests, two-sample t-tests, Wilcoxon rank-sum tests, Kruskal-Wallis tests, homogeneity of variance tests, and correlation tests, with optional group-by stratification. Also includes a p-value adjustment block for multiple comparison correction.
}
\author{
\strong{Maintainer}: Christoph Sax \email{christoph@cynkra.com} (\href{https://orcid.org/0000-0002-7192-7044}{ORCID})

Authors:
\itemize{
  \item Nicolas Bennett \email{nicolas@cynkra.com}
  \item David Granjon \email{david@cynkra.com}
  \item Mike Page \email{mike@cynkra.com}
}

Other contributors:
\itemize{
  \item  Bristol Myers Squibb [funder]
}

}
\keyword{internal}
